impl_two_traits_with_same_name
