+++ core/sr-api-macros/tests/ui/impl_two_traits_with_same_name.rs
@@ -1,0 +1,36 @@
+use runtime_primitives::traits::GetNodeBlockType;
+use test_client::runtime::Block;
+use client::{decl_runtime_apis, impl_runtime_apis};
+
+/// The declaration of the `Runtime` type and the implementation of the `GetNodeBlockType`
+/// trait are done by the `construct_runtime!` macro in a real runtime.
+struct Runtime {}
+impl GetNodeBlockType for Runtime {
+	type NodeBlock = Block;
+}
+
+decl_runtime_apis! {
+	pub trait Api {
+		fn test(data: u64);
+	}
+}
+
+mod second {
+	decl_runtime_apis! {
+		pub trait Api {
+			fn test2(data: u64);
+		}
+	}
+}
+
+impl_runtime_apis! {
+	impl self::Api<Block> for Runtime {
+		fn test(data: u64) {}
+	}
+
+	impl second::Api<Block> for Runtime {
+		fn test2(data: u64) {}
+	}
+}
+
+fn main() {}
